+++ app/src/androidTest/java/woowacourse/movie/view/MovieResultActivityTest.kt
@@ -12,6 +12,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
+import woowacourse.movie.result.view.MovieResultActivity
 
 @RunWith(AndroidJUnit4::class)
 class MovieResultActivityTest {
