+++ app/src/main/AndroidManifest.xml
@@ -13,15 +13,15 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".view.MovieResultActivity"
+            android:name=".result.view.MovieResultActivity"
             android:exported="false"
-            android:parentActivityName=".view.MovieDetailActivity"/>
+            android:parentActivityName=".detail.view.MovieDetailActivity"/>
         <activity
-            android:name=".view.MovieDetailActivity"
+            android:name=".detail.view.MovieDetailActivity"
             android:exported="false"
-            android:parentActivityName=".view.MovieMainActivity"/>
+            android:parentActivityName=".main.view.MovieMainActivity"/>
         <activity
-            android:name=".view.MovieMainActivity"
+            android:name=".main.view.MovieMainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
