+++ app/src/main/java/woowacourse/movie/main/presenter/MovieMainContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presenter
+package woowacourse.movie.main.presenter
 
 import woowacourse.movie.model.MovieAdapter
