+++ app/src/main/java/woowacourse/movie/result/presenter/MovieResultContract.kt
@@ -1,6 +1,7 @@
-package woowacourse.movie.presenter
+package woowacourse.movie.result.presenter
 
 import woowacourse.movie.model.MovieTicket
+import woowacourse.movie.presenter.BaseContract
 
 interface MovieResultContract {
     interface View : BaseContract.View {
